Refactor: Redirect 컴포넌트화
