+++ pages/redirect/index.tsx
@@ -1,25 +1,16 @@
-import { useEffect } from 'react';
-import Loading from '../../components/common/Loading';
-import UseLoginMutation from '../../state/react-query/hooks/auth';
-import { useSetRecoilState } from 'recoil';
-import { userState } from '../../state/recoil/user';
-import { LOGIN_METHOD } from '../../constants/login';
+import Head from 'next/head';
+import Redirect from '../../components/Redirect';
 
-function Redirect() {
-  const setUser = useSetRecoilState(userState);
-  const { mutate } = UseLoginMutation(LOGIN_METHOD.GITHUB);
-  useEffect(() => {
-    const code = new URL(window.location.href).searchParams.get(
-      'code'
-    ) as string;
-    mutate({ code: code });
-  }, [setUser, mutate]);
-
+const RedirectPage = () => {
   return (
     <>
-      <Loading />
+      <Head>
+        <title>로그인</title>
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <Redirect />
     </>
   );
-}
+};
 
-export default Redirect;
+export default RedirectPage;
